Crate rusoto_resourcegroupstaggingapi

Resource Groups Tagging API
This guide describes the API operations for the resource groups tagging.

A tag is a label that you assign to an AWS resource. A tag consists of a key and a value, both of which you define. For example, if you have two Amazon EC2 instances, you might assign both a tag key of "Stack." But the value of "Stack" might be "Testing" for one and "Production" for the other.

Tagging can help you organize your resources and enables you to simplify resource management, access management and cost allocation. For more information about tagging, see Working with Tag Editor and Working with Resource Groups. For more information about permissions you need to use the resource groups tagging APIs, see Obtaining Permissions for Resource Groups and Obtaining Permissions for Tagging .

You can use the resource groups tagging APIs to complete the following tasks:

Tag and untag supported resources located in the specified region for the AWS account

Use tag-based filters to search for resources located in the specified region for the AWS account

List all existing tag keys in the specified region for the AWS account

List all existing values for the specified key in the specified region for the AWS account

Not all resources can have tags. For a lists of resources that you can tag, see Supported Resources in the AWS Resource Groups and Tag Editor User Guide.

To make full use of the resource groups tagging APIs, you might need additional IAM permissions, including permission to access the resources of individual services as well as permission to view and apply tags to those resources. For more information, see Obtaining Permissions for Tagging in the AWS Resource Groups and Tag Editor User Guide.

If you're using the service, you're probably looking for ResourceGroupsTaggingApiClient and ResourceGroupsTaggingApi.

Structs

FailureInfo
Details of the common errors that all actions return.

GetResourcesInput

GetResourcesOutput

GetTagKeysInput

GetTagKeysOutput

GetTagValuesInput

GetTagValuesOutput

ResourceGroupsTaggingApiClient
A client for the AWS Resource Groups Tagging API API.

ResourceTagMapping
A list of resource ARNs and the tags (keys and values) that are associated with each.

Tag
The metadata that you apply to AWS resources to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. For more information, see Tag Basics in the Amazon EC2 User Guide for Linux Instances.

TagFilter
A list of tags (keys and values) that are used to specify the associated resources.

TagResourcesInput

TagResourcesOutput

UntagResourcesInput

UntagResourcesOutput

Enums

GetResourcesError
Errors returned by GetResources

GetTagKeysError
Errors returned by GetTagKeys

GetTagValuesError
Errors returned by GetTagValues

TagResourcesError
Errors returned by TagResources

UntagResourcesError
Errors returned by UntagResources

Traits

ResourceGroupsTaggingApi
Trait representing the capabilities of the AWS Resource Groups Tagging API API. AWS Resource Groups Tagging API clients implement this trait.