Trait rusoto_resourcegroupstaggingapi::ResourceGroupsTaggingApi

pub trait ResourceGroupsTaggingApi {
    fn get_resources(
        &self, 
        input: &GetResourcesInput
    ) -> Result<GetResourcesOutput, GetResourcesError>;

    fn get_tag_keys(
        &self, 
        input: &GetTagKeysInput
    ) -> Result<GetTagKeysOutput, GetTagKeysError>;

    fn get_tag_values(
        &self, 
        input: &GetTagValuesInput
    ) -> Result<GetTagValuesOutput, GetTagValuesError>;

    fn tag_resources(
        &self, 
        input: &TagResourcesInput
    ) -> Result<TagResourcesOutput, TagResourcesError>;

    fn untag_resources(
        &self, 
        input: &UntagResourcesInput
    ) -> Result<UntagResourcesOutput, UntagResourcesError>;
}

Trait representing the capabilities of the AWS Resource Groups Tagging API API. AWS Resource Groups Tagging API clients implement this trait.

Required Methods

fn get_resources(
&self,
input: &GetResourcesInput
) -> Result<GetResourcesOutput, GetResourcesError>

Returns all the tagged resources that are associated with the specified tags (keys and values) located in the specified region for the AWS account. The tags and the resource types that you specify in the request are known as filters. The response includes all tags that are associated with the requested resources. If no filter is provided, this action returns a paginated resource list with the associated tags.

fn get_tag_keys(
&self,
input: &GetTagKeysInput
) -> Result<GetTagKeysOutput, GetTagKeysError>

Returns all tag keys in the specified region for the AWS account.

fn get_tag_values(
&self,
input: &GetTagValuesInput
) -> Result<GetTagValuesOutput, GetTagValuesError>

Returns all tag values for the specified key in the specified region for the AWS account.

fn tag_resources(
&self,
input: &TagResourcesInput
) -> Result<TagResourcesOutput, TagResourcesError>

Applies one or more tags to the specified resources. Note the following:

Not all resources can have tags. For a list of resources that support tagging, see Supported Resources in the AWS Resource Groups and Tag Editor User Guide.

Each resource can have up to 50 tags. For other limits, see Tag Restrictions in the Amazon EC2 User Guide for Linux Instances.

You can only tag resources that are located in the specified region for the AWS account.

To add tags to a resource, you need the necessary permissions for the service that the resource belongs to as well as permissions for adding tags. For more information, see Obtaining Permissions for Tagging in the AWS Resource Groups and Tag Editor User Guide.

fn untag_resources(
&self,
input: &UntagResourcesInput
) -> Result<UntagResourcesOutput, UntagResourcesError>

Removes the specified tags from the specified resources. When you specify a tag key, the action removes both that key and its associated value. The operation succeeds even if you attempt to remove tags from a resource that were already removed. Note the following:

To remove tags from a resource, you need the necessary permissions for the service that the resource belongs to as well as permissions for removing tags. For more information, see Obtaining Permissions for Tagging in the AWS Resource Groups and Tag Editor User Guide.

You can only tag resources that are located in the specified region for the AWS account.