Struct rusoto_resourcegroupstaggingapi::ResourceGroupsTaggingApiClient

pub struct ResourceGroupsTaggingApiClient { /* fields omitted */ }

A client for the AWS Resource Groups Tagging API API.

Methods

impl ResourceGroupsTaggingApiClient

pub fn new(region: Region) -> ResourceGroupsTaggingApiClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ResourceGroupsTaggingApiClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,